icons, navbar update
