+++ GnuOne/ClientApp/src/Pages/MyProfile.js
@@ -2,7 +2,7 @@
 import { Link, Switch } from 'react-router-dom';
 import friends from '../icons/friends.svg'
 import ProfileWheel from '../components/NewComponents/ProfileWheel';
-import Navbar from '../components/NewComponents/NavBar';
+import Navbar from '../components/NewComponents/Navbar/NavBar';
 
 
 //testar
