+++ GnuOne/ClientApp/src/components/NewComponents/Navbar/navbar.css
@@ -1,0 +1,6 @@
+.navbar {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    padding: 1em;
+}
